Add soft delete and improve service management

+++ src/main/java/com/jonathangomz/economy21/repository/ServicesRepository.java
@@ -2,5 +2,5 @@ package com.jonathangomz.economy21.repository;
 
 import com.jonathangomz.economy21.model.Service;
 
-public interface ServicesRepository extends CrudRepository<Service, Integer> {
+public interface ServicesRepository extends CrudRepository<Service, Long> {
 }
